feat:add date and temperature to trip details page
