+++ pages/trip/[id].js
@@ -1,6 +1,7 @@
 import { useRouter } from "next/router";
 import useSWR from "swr";
 import TripDetails from "@/components/TripDetails";
+import WeatherList from "@/components/WeatherList";
 
 const fetcher = (...args) => fetch(...args).then((res) => res.json());
 
@@ -28,15 +29,18 @@ export default function TripDetailsPage() {
   const endDate = timePeriod.map((timePeriod) => `${timePeriod.endDate}`);
 
   return (
-    <TripDetails
-      title={title}
-      location={location}
-      handleDelete={handleDelete}
-      img={img}
-      timePeriod={timePeriod}
-      id={id}
-      startDate={startDate}
-      endDate={endDate}
-    />
+    <>
+      <TripDetails
+        title={title}
+        location={location}
+        handleDelete={handleDelete}
+        img={img}
+        timePeriod={timePeriod}
+        id={id}
+        startDate={startDate}
+        endDate={endDate}
+      />
+      <WeatherList />
+    </>
   );
 }
